refactor: move update alias functionality to trait
